+++ source/pipes/date-time-to-utc.ts
@@ -1,0 +1,13 @@
+import { Pipe, PipeTransform } from "@angular/core";
+import { DateTime } from "luxon";
+
+@Pipe({
+  name: "dateTimeToUtc"
+})
+export class DateTimeToUtcPipe implements PipeTransform {
+
+  transform(value: DateTime): DateTime {
+    return value.toUTC();
+  }
+
+};